+++ src/CECodeHelper/CECode.TeamCity/Service/ITeamCityService.cs
@@ -8,11 +8,11 @@ namespace CECode.TeamCity.Service
         Build GetAdvantageBuild();
         Build GetAdvantagePatches();
         Build GetBuildByType(string buildType);
-        Build GetBuildById(int id);
+        Build GetBuildById(long id);
         Build GetBuildByNumber(string number);
         IList<RunningBuild.Build> GetRunningBuilds();
         IList<RunningBuild.Build> GetBuilds(string locator);
         BuildDetails GetBuildDetails(Build build);
-        BuildDetails GetBuildDetails(int id);
+        BuildDetails GetBuildDetails(long id);
     }
 }
